requirements.txt
beautifulsoup4==4.14.2
fastapi==0.119.0
pinecone==7.3.0
pydantic==2.12.3
python-dotenv==1.1.1
Requests==2.32.5
selenium==4.37.0
tqdm==4.67.1
uvicorn==0.38.0
webdriver_manager==4.0.2
